+++ src/database/pipelines/rawVoteCsvParser.ts
